pppoe: T4391: bugfix IPv6 DHCP-PD not working after reboot
When VyOS is booting and an interface is brought up (PPPoE) which requires a
user callback script that is executed asynchronously when the interface is up
we can not use Config(). The problem is, Config() is not available when
the system starts and the initial commit is still processed.

We need to move to ConfigTreeQuery() which was build for this exact same
purpose.