bond: T4522: add ability to specify mii monitor interval via CLI
Linux Kernel supports to specify the MII link monitoring frequency in
milliseconds. This determines how often the link state of each slave is
inspected for link failures. A value of zero disables MII link monitoring.
A value of 100 is a good starting point.

The default value is 100.

set interfaces bonding bond0 mii-mon-interval <n>