bgp: T4257: Changing BGP "local-as" to "system-as"
bgp: T4257: Changing BGP "local-as" to "system-as"

This change is to change the global BGP name for the node "local-as" to "system-as"

This is needed so that there's less ambiguity with the local-as feature per neighbor.

bgp: T4257: Changing BGP "local-as" to "system-as"

bgp: T4257: Changing BGP "local-as" to "system-as"

This change is to change the global BGP name for the node "local-as" to "system-as"

This is needed so that there's less ambiguity with the local-as feature per neighbor.