T4550: router-advert: Add deprecate-prefix & decrement-lifetimes options
DeprecatePrefix and DecrementLifetimes options in radvd is useful in a DHCPv6-PD environment to accommodate prefix changes from ISP's delegating router. Though there is currently no integration between the DHCP PD client (wide-dhcpv6-client) and radvd, it could be a good start point to have the 2 options configurable by the user.

https://phabricator.vyos.net/T4550

- deprecate-prefix: Upon shutdown, deprecate the prefix. This is useful
  in a DHCPv6 PD environment: When ISP re-assigns a new prefix,
deprecate the old prefix that was advertised.
- decrement-lifetimes: Decrement the values of the preferred and valid
  lifetimes for the prefix over time. This is also useful in a DHCPv6
  PD environment to keep the advertised prefix's lifetimes in sync with
  the prefix from delegating router.