Update protocols_isis.py
isis: T4693: Fix ISIS segment routing configurations

This change is to fix more bugs in which ISIS segment routing was broken due to a refactor. This change also introduces a few additions to the ISIS handler for checking per prefix validations for segment value and mutual exclusivity for two options.