wireguard: T4702: actively revoke peer if it gets disabled
When any configured peer is set to `disable` while the Wireguard tunnel is up
and running it does not get actively revoked and removed. This poses a security
risk as connections keep beeing alive.

Whenever any parameter of a peer changes we actively remove the peer and fully
recreate it on the fly.