T4496: Added lists of values in the help of op-mode ping command
Added list of possible VRFs in the help of the ping command
Added list of possible interfaces in the help of the ping command
Changed, if an option was selected before in the ping command,
it does not appear in possible completion.
Added error message when an unexpected option was selected.