openvpn: T3214: specify nopool on --server line only if needed
The --server directive will already create a pool automatically.
For this reason noppol should be used only when an explicit
client-ip-pool was configured by the user.

If that's not the case, then the nopool flag should not be
specified and no manual pool should be configured.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>
(cherry picked from commit 0ccbbca01b22232b5cba63d64ab00eb54af7b068)