op-mode: dns-forwarding: T4578: drop sudo calls
Commit 66288ccfee ("dns-forwarding: T4578: Rewrite show dns forwarding") added
the implementation for the new standardized op-mode definitions/implementation.

As the API daemon has the proper permissions and also the CLI op-mode calls the
script already with "sudo", there is no need to call "sudo" inside this script,
again.

Also add dns.py to data/op-mode-standardized.json for the GraphQL schema to be
generated.