T4865: Fix to generate container image from the file
In case if we want generate own container image from a Dockerfile
and if it requires update or install packages in container we get
error. As it tries to use default network 'podman' and do own NAT
translations via 'iptables'. If fact we don't use iptables in 1.4
As result it cannot build such image.
Use '--net host' to fix it.