op-mode: T4952: improve interface completion helper CLI experience
There are currently two ways to generate the interface name completion helper
list (we use openvpn in this example)

- <script> ${vyos_completion_dir}/list_interfaces.py --type openvpn</script>
- <path>interfaces openvpn</path>

The first one using <script> tends to be rather slow as there is a Python
interpreter startup involved (expensive). The latter simply calls a C program
which is executed rather fast and gives the same result.

We can simply replace the first call with the second one to make the CLI
feel faster.