bgp: T4817: adjust CLI definition for local-role
Instead of making local-role a tagNode - which might look suitable in the
first place - we change it to a leafNode.

Using a tagNode will suggest the user can enter "any data" indicated by > at
the end.

vyos@vyos# set protocols bgp neighbor 1.1.1.1 local-role
Possible completions:
 > customer             Using Transit
 > peer                 Public/Private Peering
 > provider             Providing Transit
 > rs-client            RS Client
 > rs-server            Route Server
 >

Moving to a leafNode will only five the user discrete choiced. The downside is
that the require an additional leafNode for the strict parameter.