snmp: T4857: explicitly define default community networks 0.0.0.0/0 and ::/0
After the RESTRICTED view was introduced snmpd requires a network to be
specified. Before adding the RESTRICTED view snmpd always assumed the default
network 0.0.0.0/0.

This commit re-adds the build in default networks for IPv4 and IPv6 and
exposes it as a proper default to the CLI so the user is informed about it:

vyos@vyos# set service snmp community foooo
Possible completions:
   authorization        Authorization type (default: ro)
+  client               IP address of SNMP client allowed to contact system
+  network              Subnet of SNMP client(s) allowed to contact system (default:
                        0.0.0.0/0, ::/0)