T4997: add dhcp client user hooks dir
T4997: add dhcp client user hooks dir
This commit adds a script to run user-defined hook scripts upon renewing
a DHCP lease. This can be used to, for example, dynamically define a
firewall address-group based on the dynamic IP address of an interface.

For an example of its use (as well as the use case I had in mind while
coding this), see https://vyos.dev/T2196#142394