login: T5039: Support hashing rounds in `encrypted-password` values
Since glibc 2.7, the SHA-256 and SHA-512 implementations support
a user-supplied number of hashing rounds, defaulting to 5000.  If
the "$id$" characters in the salt are followed by "rounds=xxx$",
where xxx is an integer, then the result has the form

      $id$rounds=yyy$salt$encrypted

where yyy is the number of hashing rounds actually used.  The
number of rounds actually used is 1000 if xxx is less than 1000,
999999999 if xxx is greater than 999999999, and is equal to xxx
otherwise.