ethernet: T3891: add conditional code-path when doing speed/duplex changes
There is no need for the backend code to call ethtool and try to change speed or
duplex settings every time there is a change in the interface configuration,
but no change for the speed/duplex subnodes. This also makes the commit itself
faster when working with ethernet interfaces.

Bonus: no repeating CLI messages that the driver does not support speed/duplex
changes, as we do not change anything here.

Extension to commit f2ecc9710 ("ethernet: T3891: honor auto-negotiation support
per NIC")