T5295: Fix QoS shaper rate limit
Do not handle rate via 'tc filter' directly but rather set the
'tc filter' to direct traffic to the correct tc class flow.
As it in 1.3.
It fixes random unexpected shapes, when you set for example 300mbit
but get 3-11mbit

Current implementation seems not correct as it uses rate limits
two times (in class and in filter):
  tc class replace dev eth0 parent 1:1 classid 1:17 htb rate 250000000 \
      burst 15k quantum 1514
  tc filter replace dev eth0 parent 1: protocol all u32 match \
      ip dst 192.168.122.11 action police rate 250000000 burst 15k flowid 1:17

The correct way after fix:
  tc class replace dev eth0 parent 1:1 classid 1:17 htb rate 250000000 \
      burst 15k quantum 1514
  tc filter replace dev eth0 parent 1: protocol all u32 match \
      ip dst 192.168.122.11 flowid 1:17