T5533: Fix VRRP IPv6 FAULT state due to IPv6 tentative state
Checks if an IPv6 address on a specific network interface is
in the tentative state. IPv6 tentative addresses are not fully configured
and are undergoing Duplicate Address Detection (DAD) to ensure they are
unique on the network.

inet6 2001:db8::3/125 scope global tentative

It tentative state the group enters in FAULT state. Fix it.