ddclient: T5612: Improve dual stack support for dyndns2 protocol
dyndns2 protocol in ddclient honors dual stack for selective servers
because of the way it is implemented in ddclient.

We formalize the well known servers that support dual stack in a list
and check against it when validating the configuration.

(cherry picked from commit 98694f5717f282e10caeded3e4e07891fd5bea18)