init: T5239: configure system hostname prior to FRR startup
On first boot after an upgrade /etc/hostname and FRR configuration is not
populated. FRR determines the system hostname once during startup and does not
repect changes of the hostname CLI value.

Thus after an upgrade of VyOS FRR started with a hostname of debian that was
propagated to peers.

The commit retrieves the hostname from the CLI and presets this before FRR is
initially started.

(cherry picked from commit ac21a4e69fac27504b62927a20d0a6a273abb034)