pppoe: T5630: make MRU default to MTU if unspecified
This fixes the implementation in e062a8c11 ("pppoe: T5630: allow to specify MRU
in addition to already configurable MTU") and restores the bahavior that MRU
defaults to MTU if MRU is not explicitly set.

This was the behavior in VyOS 1.3.3 and below before we added ability to define
the MRU value.

(cherry picked from commit ffd7339e2ea3eafdd97ac0763ca4a3913fe71bf3)