ddclient: T5791: Update dynamic dns configuration path
Modify the configuration path to be consistent with the usual dialects
of VyoS configuration (wireguard, dns, firewall, etc.)

This would also shorten the configuration path and have a unified
treatment for RFC2136-based updates and other 'web-service' based updates.

While at it, add support for per-service web-options. This would allow
for probing different external URLs on a per-service basis.