frr: T4020: add option to define number of open file descriptors
This allows the operator to control the number of open file descriptors each
daemon is allowed to start with. The current assumed value on most operating
systems is 1024.

If the operator plans to run bgp with several thousands of peers then this is
where we would modify FRR to allow this to happen.

set system frr descriptors <n>

(cherry picked from commit 892c28ccf634173d4c4952c248cb03974c560793)