dhcp: T5846: Ensure DUID regex range is bound
The DUID regex was missing a lower bound, which could cause it not to
match when it should.

We have to specify the lower bound explicitly as 0 to keep the regex
behavior similar to that in Python (in Python, omitting the lower bound
is equivalent to specifying 0).

(cherry picked from commit 551f06218755076cde588c848c01ce5ca1bf5e6b)