login: T5875: restore home directory permissions when re-adding user account
After deleting a user account and working with a newly added account, we see
that after rebooting in the previously saved configuration, the user is
re-added but it's home directory might have an old UID set on the filesystem.

This is due to the fact that vyos config does not store UIDs. When adding a
user account to the system we now check if the home directory already exists
and adjust the ownership to the new UID.

(cherry picked from commit 3c990f49e2bf9347bd2cc478995baa995ee822fd)