login: T5875: restore home directory permissions only when needed
This improves commit 3c990f49e ("login: T5875: restore home directory
permissions when re-adding user account") in a way that the home directory
owner is only altered if it differs from the expected owner.

Without this change on every boot we would alter the owner which could increase
the boot time if the home of a user is cluttered.