smoketest: T5195: fix BasicInterfaceTest tearDown() timeout penalty
Commit ad9bdfc24 ("T5195: add timeout argument to process_named_running()")
added a 2*10 seconds penalty for every interface test (dhcp and dhcpv6).
This leads to long runs of "make test" after an ISO build.

There is no need to wait 10 seconds for a test that checks for a process
not running. The timeout is there to give the process some time to startup.

(cherry picked from commit 041db49533d57cabfccd319492b85ee0bafdd40c)